Rename AJ job class as there is only one now
